Crate rust_keylock

The rust-keylock library

Executes the logic of the rust-keylock.

This library is the executor of the rust-keylock logic. Editor references are used to interact with the rust-keylock users.

Modules

datacrypt
Defines the Cryptor trait and implements the encryption and decryption for the rust-keylock

Structs

Entry
Struct that defines a password entry.

Props
A struct that allows storing general configuration values. The configuration values are stored in plaintext.

Safe
Holds the data that should be safe and secret.

Enums

Menu
Enumeration of the several different Menus that an Editor implementation should handle.

UserSelection
Represents a User selection that is returned after showing a Menu.

Traits

Editor
Trait to be implemented by various different Editors (Shell, Web, Android, other...).

Functions

execute
Takes a reference of Editor implementation as argument and executes the rust-keylock logic. The Editor is responsible for the interaction with the user. Currently there are Editor implementations for shell and for Android.