Struct rust_keylock::datacrypt::EntryPasswordCryptor

pub struct EntryPasswordCryptor { /* fields omitted */ }

Encrypts and decrypts passwords of Entries in order not to be kept in the memory in plain.

Methods

impl EntryPasswordCryptor

fn new() -> EntryPasswordCryptor

Creates a new EntryPasswordCryptor.

fn encrypt_str(&self, input: &str) -> Result<String, RustKeylockError>

Gets a String input and returns it encrypted and Base64-encoded

fn decrypt_str(&self, input: &str) -> Result<String, RustKeylockError>

Gets a Base64-encoded String input and returns it decrypted