Enum rust_keylock::UserSelection

pub enum UserSelection {
    NewEntry(Entry),
    ReplaceEntry(usize, Entry),
    DeleteEntry(usize),
    GoTo(Menu),
    ProvidedPassword(String, usize),
    Ack,
    ExportTo(String),
    ImportFrom(String, String, usize),
}

Represents a User selection that is returned after showing a Menu.

Variants
NewEntry(Entry)

The User selected an Entry.

ReplaceEntry(usize, Entry)

The User updated an Entry.

DeleteEntry(usize)

The User deleted an Entry.

GoTo(Menu)

The User selected to go to a Menu.

ProvidedPassword(String, usize)

The User provided a password and a number.

Ack

The User acknowledges something.

ExportTo(String)

The User selected to export the password Entries to a path.

ImportFrom(String, String, usize)

The User selected to import the password Entries from a path.