Trait rust_keylock::Editor

pub trait Editor {
    fn show_password_enter(&self) -> UserSelection;
    fn show_change_password(&self) -> UserSelection;
    fn show_menu(&self, menu: &Menu, entries: &[Entry]) -> UserSelection;
    fn exit(&self, contents_changed: bool) -> UserSelection;
    fn show_message(&self, message: &'static str) -> UserSelection;

    fn sort_entries(&self, entries: &mut [Entry]) { ... }
}

Trait to be implemented by various different Editors (Shell, Web, Android, other...).

It drives the interaction with the Users

Required Methods

fn show_password_enter(&self) -> UserSelection

Shows the interface for entering a Password and a Number.

fn show_change_password(&self) -> UserSelection

Shows the interface for changing a Password and/or a Number.

fn show_menu(&self, menu: &Menu, entries: &[Entry]) -> UserSelection

Shows the specified Menu to the User.

fn exit(&self, contents_changed: bool) -> UserSelection

Shows the Exit Menu to the User.

fn show_message(&self, message: &'static str) -> UserSelection

Shows a message to the User.

Provided Methods

fn sort_entries(&self, entries: &mut [Entry])

Sorts the supplied entries.