Crate rust_keylock

The rust-keylock library

Executes the logic of the rust-keylock.

This library is the executor of the rust-keylock logic. Editor references are used to interact with the rust-keylock users.

How it works

The library implements the functionality where a User stores password Entries in an encrypted toml file. Currently, AES with CTR is used and the encryption key is generated with Bcrypt, but more choices can be implemented in the future.

The iv that is used for encyption/decryption and the salt for the generation of the encryption key are stored along with the rest of the encrypted data bytes in the same file.

The 16-byte iv is always in the start of the file.

The 16 bytes of salt are merged along with the encrypted actual data. The position of the salt is actually defined by the user. Upon rust-keylock initialization, the user sets up two things:

A master password, which is combined with an initially pseudo-randomly generated 16-bytes array, the salt, in order to create the Bcrypt key that is used for the data encryption/decryption.

A number, which is the actual location of the salt among the encrypted data.

Having the salt inbetween the actual encrypted data gives two obvious benefits:

Makes the encrypted file portable

All the information that is needed for encryption/decryption is in the file itself. Nothing additional is needed. Therefore, the encrypted file can just be copied in some location to be backed-up, or even be synchronized in other devices and be used there as well.

Increases the security of the encryption

The actual data bytes are "interrupted" by the salt and this potentially makes an adversary's job more difficult; more difficult to infer things about the data contents or the encryption key.

A concrete example

Let's assume that the user has set his master password and the number 3 as the additional number to the password.

Let's also assume that the actual password Entries data gets encrypted as the following byte array: 0x11u8, 0x12u8, 0x13u8, 0x14u8, 0x15u8, 0x16u8, 0x17u8, 0x18u8, 0x19u8, 0x20u8, 0x21u8, 0x22u8, 0x23u8, 0x24u8, 0x25u8, 0x26u8, 0x27u8, 0x28u8, 0x29u8, 0x30u8, 0x31u8, 0x32u8, 0x33u8, 0x34u8, 0x35u8, 0x36u8, 0x37u8, 0x38u8, 0x39u8, 0x40u8, 0x41u8, 0x42u8

Upon saving to the file, the rust-keylock will generate two pseudo-random byte arrays; one to be the iv and one to be the salt. Let the iv be 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8.

Let the salt be 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8.

All that makes the actual bytes that are saved in the file to be:

The iv

Three bytes of real encrypted data

The salt

The rest of the real encrypted data bytes

Thus, the saved data should be looking like following:

0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x01u8, 0x11u8, 0x12u8, 0x13u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x02u8, 0x14u8, 0x15u8, 0x16u8, 0x17u8, 0x18u8, 0x19u8, 0x20u8, 0x21u8, 0x22u8, 0x23u8, 0x24u8, 0x25u8, 0x26u8, 0x27u8, 0x28u8, 0x29u8, 0x30u8, 0x31u8, 0x32u8, 0x33u8, 0x34u8, 0x35u8, 0x36u8, 0x37u8, 0x38u8, 0x39u8, 0x40u8, 0x41u8, 0x42u8

Modules

datacrypt
Defines the Cryptor trait and implements the encryption and decryption for the rust-keylock

Structs

Entry
Struct that defines a password entry.

Props
A struct that allows storing general configuration values.

Enums

Menu
Enumeration of the several different Menus that an Editor implementation should handle.

UserSelection
Represents a User selection that is returned after showing a Menu.

Traits

Editor
Trait to be implemented by various different Editors (Shell, Web, Android, other...).

Functions

execute
Takes a reference of Editor implementation as argument and executes the rust-keylock logic. The Editor is responsible for the interaction with the user. Currently there are Editor implementations for shell and for Android.